Factorize NotFound component into StatusPage
